Fix steps.stories.js - overview hooks issue
